♻️ refactor: 마이그레이션 기능 및 함수를 분리 및 리팩터링
- 저장을 직접 진행하는 마이그레이션 함수와 다음 버전의 데이터로 반환된 값만 반환하는 함수들을 분리
- 테스트하기에도 용이
